Create html-basic-skele.html
This is literally just a super basic HTML document that you can use for whatever it has header, main, footer literally so simplistic and on the footer it has a link that you can put whatever to like back to main or where ever you would want it to go to. I wanted to upload this here because of its simplicity.